Refactored input_generator to prognostic run
